rtt: virtualization and switch to 32khz
I tried to switch the rtt to 32khz resolution, and then virtualize its alarm such that we can derive a new TMilli from one of the virtualized alarms. However, this doesn't seem to work for some reason or the other. Debugging it revealed that the code works while stepping through it in Blink (i.e., the LEDs toggle). However, once running at full speed, and interrupting the code again removes the LED tasks from the scheduler. No idea why though.